added choice on commit to master or deploy tag

+++ commit.sh
@@ -25,13 +25,13 @@ components['plinker-test']='plinker/test'
 increment_semver() {
     # passed in as: -p, -m, -M | release = p, m, M
     release=${1#?}
-    
-    # passed in as: v0.0.1 | version = 0.0.1 
+
+    # passed in as: v0.0.1 | version = 0.0.1
     version=${2#?}
-    
+
     # explode into array based on .
     v=(${version//./ })
-    
+
     # default to 0.0.1
     if [ ${#v[@]} -ne 3 ]
     then
@@ -49,12 +49,12 @@ increment_semver() {
             ((v[1]++))
               v[2]=0
         fi
-        
+
         # Patch
         if [ "$release" == "p" ]; then
             ((v[2]++))
         fi
-        
+
         echo "${v[0]}.${v[1]}.${v[2]}"
     fi
 }
@@ -62,9 +62,35 @@ increment_semver() {
 #
 commit() {
 
+    echo "- Fetching current changes"
+    git fetch --all
+
+    # get the current semantic tag
+    latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
+    echo "Current Semantic Version Tag: $latestTag"
+
+    # stage changes
+    git add -A ./
+
+    # stage changes
+    git commit -a -m "[since: v$latestTag]  $date - $message"
+
+    # pull latest
+    git pull origin master
+
+    # commit any remote changes
+    git commit -a -m  "[since: v$latestTag]  $date - $message"
+
+    # update master
+    git push origin master
+}
+
+#
+deploy_tag() {
+
     echo "- Fetching tags"
     git fetch --tags
-    
+
     # get the current semantic tag
     latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
     echo "Current Semantic Version Tag: $latestTag"
@@ -72,7 +98,7 @@ commit() {
     # look for previous semvar choice tmp file
     if [ ! -f /tmp/semvar_choice ]; then
         while true; do
-            read -p "Which type of change has been done: [p]atch, [m]inor, [M]ajor? " semver
+            read -p "Which type of changes has been done: [p]atch, [m]inor, [M]ajor? " semver
             case $semver in
                 [p]* ) break;;
                 [m]* ) break;;
@@ -82,31 +108,31 @@ commit() {
         done
         echo $semver > /tmp/semvar_choice
     else
-        # red choice into semvar variable
+        # read choice into semvar variable
         read -r semver < /tmp/semvar_choice;
     fi
-    
-    # new release semvar 
+
+    # new release semvar
     releaseSemvar=$(increment_semver -"$semver" $latestTag)
 
     # stage changes
     git add -A ./
-    
+
     # stage changes
     git commit -a -m "[since: v$releaseSemvar]  $date - $message"
-    
+
     # pull latest
     git pull origin master
-    
+
     # commit any remote changes
     git commit -a -m  "[since: v$releaseSemvar]  $date - $message"
-    
+
     # update master
     git push origin master
-    
+
     # commit tag
     git tag -a v$releaseSemvar -m "[$semver] v$releaseSemvar - $date - $message"
-    
+
     # push tag
     git push origin v$releaseSemvar
 }
@@ -120,6 +146,17 @@ main() {
     echo "- Message -: $message"
     echo "---------------------------------------------------"
 
+    # ask what to do, push to master or deploy to master and tag
+    while true; do
+        read -p "Do you wish to [c]ommit, [d]eploy or [e]xit? " dep
+        case $dep in
+            [c]* ) break;;
+            [d]* ) break;;
+            [e]* ) exit;;
+            * ) echo "Please answer c, d or e.";;
+        esac
+    done
+
     # move into project workspace
     cd $projectDir
 
@@ -127,18 +164,24 @@ main() {
     for key in "${!components[@]}"
     do
         echo "- Commiting: $key"
-        
+
         # check project folder exists
         if [ -d "$PWD/vendor/${components[$key]}" ]; then
-            
+
             if [ -d "$PWD/vendor/${components[$key]}/.git" ]; then
-            
+
                 echo "- Entering: $PWD/vendor/${components[$key]}"
                 cd "$PWD/vendor/${components[$key]}"
-                
+
                 # commit
-                commit $message
-            
+                if [ "$dep" == "c" ]; then
+                    commit $message
+                fi
+
+                # deploy tag
+                if [ "$dep" == "d" ]; then
+                    deploy_tag $message
+                fi
             else
                 echo "- Skipping, .git directory does not exist: $PWD/vendor/${components[$key]}/.git"
             fi
@@ -148,11 +191,14 @@ main() {
 
         echo "---------------------------------------------------"
 
-        # move back into project workspace        
+        # move back into project workspace
         cd $projectDir
     done
-    
-    rm /tmp/semvar_choice
+
+    # remove semvar choice tmp file
+    if [ -f /tmp/semvar_choice ]; then
+        rm /tmp/semvar_choice
+    fi
 }
 
 main
